requirements.txt
pandas
sqlalchemy
psycopg2-binary
polars
selenium
beautifulsoup4
requests
webdriver-manager
apache-airflow